pyproject.toml
[tool.poetry]
name = "aithena-services"
version = "0.1.0"
description = ""
authors = ["Camilo Velez <camilo.velez@axleinfo.com>, Antoine Gerardin <antoine.gerardin@nextonicsolutions.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.4"
openai = "^1.35.13"
requests = "^2.32.3"
ollama = "^0.3.1"
llama-index-llms-ollama = "^0.2.2"
llama-index = "==0.10.62"
llama-index-llms-openai = "^0.1.29"
llama-index-embeddings-ollama = "^0.1.3"
fastapi = {extras = ["standard"], version = "^0.112.0"}
llama-index-embeddings-azure-openai = "^0.1.11"
pytest-asyncio = "^0.23.8"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
#log_file = "pytest.log"
#log_file_level = "INFO"
#log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
#log_file_date_format = "%Y-%m-%d %H:%M:%S"
